More counting code

+++ HullSolver.hpp
@@ -47,7 +47,7 @@ signals:
   void hullPointFound(const int& x, const int& y) const;
 
   void solutionFound(const HullTimeline&) const;
-  void solutionFound(const QString& a, const int& b, const int& c) const;
+  void solutionFound(const QString& a, const int& b, const double& c) const;
 
   void algorithmChanged(const QString&) const;
   void inputChanged(const QString&) const;
